OmniWrenchComponent修正
