+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/Component/Player/PlayerJumpAttackComponent.cpp
@@ -1,6 +1,6 @@
 #include "PlayerJumpAttackComponent.h"
 
-#include "PlayerWeaponComponent.h"
+#include "PlayerOmniWrenchComponent.h"
 
 
 my::PlayerJumpAttackComponent::PlayerJumpAttackComponent(int priority) :
@@ -28,7 +28,7 @@ std::string_view my::PlayerJumpAttackComponent::GetStateType(void) const {
 
 bool my::PlayerJumpAttackComponent::Initialize(void) {
     super::Initialize();
-    _weapon_com = super::GetOwner()->GetComponent<my::PlayerWeaponComponent>();
+    _weapon_com = super::GetOwner()->GetComponent<my::PlayerOmniWrenchComponent>();
     return true;
 }
