+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/Component/Player/PlayerThrowAttackEndComponent.h
@@ -12,7 +12,7 @@ class PlayerThrowAttackEndComponent : public my::PlayerActionComponent {
     using super = my::PlayerActionComponent;
 private:
     //! 武器
-    std::weak_ptr<class PlayerWeaponComponent> _weapon_com;
+    std::weak_ptr<class PlayerOmniWrenchComponent> _weapon_com;
 public:
     /// <summary>
     /// コンストラクタ
