--- ClientProject/Mof/RachetNClankCopy/RachetApplication/Component/Weapon/OmniWrenchDefaultComponent.h
@@ -1,88 +1,0 @@
-#ifndef MY_OMNI_WRENCH_DEFAULT_COMPONENT_H
-#define MY_OMNI_WRENCH_DEFAULT_COMPONENT_H
-
-
-#include "../ActionComponent.h"
-
-#include <memory>
-
-
-namespace my {
-class OmniWrenchDefaultComponent : public my::ActionComponent {
-    using super = my::ActionComponent;
-private:
-    //! 速度
-    std::weak_ptr<class VelocityComponent> _velocity_com;
-    //! 状態
-    std::weak_ptr<class OmniWrenchActionStateComponent > _action_state_com;
-    /// <summary>
-    /// 変更
-    /// </summary>
-    /// <param name="name"></param>
-    void ChageState(const std::string& name);
-public:
-    /// <summary>
-    /// コンストラクタ
-    /// </summary>
-    /// <param name="priority"></param>
-    OmniWrenchDefaultComponent(int priority);
-    /// <summary>
-    /// コピーコンストラクタ
-    /// </summary>
-    /// <param name="obj"></param>
-    OmniWrenchDefaultComponent(const OmniWrenchDefaultComponent& obj);
-    /// <summary>
-    /// デストラクタ
-    /// </summary>
-    virtual ~OmniWrenchDefaultComponent();
-    /// <summary>
-    /// ゲッター
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual std::string GetType(void) const override;
-    /// <summary>
-    /// ゲッター
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual std::string_view GetStateType(void) const override;
-    /// <summary>
-    /// 初期化
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual bool Initialize(void) override;
-    /// <summary>
-    /// 更新
-    /// </summary>
-    /// <param name="delta_time">時間</param>
-    /// <returns></returns>
-    virtual bool Update(float delta_time) override;
-    /// <summary>
-    /// 解放
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual bool Release(void) override;
-    /// <summary>
-    /// 複製
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual std::shared_ptr<my::Component> Clone(void) override;
-    /// <summary>
-    /// 開始
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns>成功</returns>
-    virtual bool Start(void) override;
-    /// <summary>
-    /// 終了
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns>成功</returns>
-    virtual bool End(void) override;
-};
-}
-#endif // !MY_OMNI_WRENCH_DEFAULT_COMPONENT_H
