+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/Factory/ComponentRegister/OmniWrenchComponentRegister.h
@@ -1,0 +1,3 @@
+my::ComponentFactory::AddComponentCreator<my::OmniWrenchActionStateComponent>();
+my::ComponentFactory::AddComponentCreator<my::OmniWrenchDefaultComponent>();
+my::ComponentFactory::AddComponentCreator<my::OmniWrenchThrowedComponent>();
