+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/RachetApplication.vcxproj.filters
@@ -237,12 +237,6 @@
     <Filter Include="Rachet\Source\Factory\Builder">
       <UniqueIdentifier>{01534295-06e6-4519-a38c-12387d9b3c88}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Rachet\Source\Component\Weapon">
-      <UniqueIdentifier>{e65bf001-c5e9-4aae-ba98-462ca64d800d}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Rachet\Header\Component\Weapon">
-      <UniqueIdentifier>{94cb0e0c-0632-41b0-b66f-ad29f47dffb4}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Rachet\Header\Factory\ComponentRegister">
       <UniqueIdentifier>{e9111719-68d3-4b4e-8894-fe7bdacd2cb1}</UniqueIdentifier>
     </Filter>
@@ -273,6 +267,24 @@
     </Filter>
     <Filter Include="Resource\stage">
       <UniqueIdentifier>{492cf9dc-39e2-4b04-9c3a-8bc9da18b19f}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Rachet\Header\Component\Weapon">
+      <UniqueIdentifier>{3c30a394-d063-4f3d-afe5-6d70cd7d2a89}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Rachet\Header\Component\Weapon\OmniWrench">
+      <UniqueIdentifier>{20c591a8-4d3e-4996-a818-959ba0fb102b}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Rachet\Header\Component\Weapon\Blaster">
+      <UniqueIdentifier>{deff356c-adb3-4d9d-ae53-ffbc3d74fad5}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Rachet\Source\Component\Weapon">
+      <UniqueIdentifier>{3964a09b-0c7a-47a3-9cd3-ab68c78cc300}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Rachet\Source\Component\Weapon\OmniWrench">
+      <UniqueIdentifier>{d2d7b52e-e01f-41b1-8910-3bde8bc2b663}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Rachet\Source\Component\Weapon\Blaster">
+      <UniqueIdentifier>{1e583bb3-3b36-443a-8c30-028c5cbf7660}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -720,20 +732,11 @@
     <ClCompile Include="Component\Collision\Algolithm\EnemyOmniWrenchCollisionAlgolithm.cpp">
       <Filter>Rachet\Source\Component\Collision\Algolithm</Filter>
     </ClCompile>
-    <ClCompile Include="Component\Player\PlayerWeaponComponent.cpp">
+    <ClCompile Include="Component\Player\PlayerOmniWrenchComponent.cpp">
       <Filter>Rachet\Source\Component\Player</Filter>
-    </ClCompile>
-    <ClCompile Include="Component\Weapon\OmniWrenchThrowedComponent.cpp">
-      <Filter>Rachet\Source\Component\Weapon</Filter>
-    </ClCompile>
-    <ClCompile Include="Component\Weapon\OmniWrenchActionStateComponent.cpp">
-      <Filter>Rachet\Source\Component\Weapon</Filter>
     </ClCompile>
     <ClCompile Include="State\OmniWrenchActionStateImplementation.cpp">
       <Filter>Rachet\Source\State</Filter>
-    </ClCompile>
-    <ClCompile Include="Component\Weapon\OmniWrenchDefaultComponent.cpp">
-      <Filter>Rachet\Source\Component\Weapon</Filter>
     </ClCompile>
     <ClCompile Include="Camera\FirstPersonCameraController.cpp">
       <Filter>Rachet\Source\Camera</Filter>
@@ -809,6 +812,15 @@
     </ClCompile>
     <ClCompile Include="Component\Player\PlayerThrowAttackSetComponent.cpp">
       <Filter>Rachet\Source\Component\Player\Action</Filter>
+    </ClCompile>
+    <ClCompile Include="Component\Weapon\OmniWrench\OmniWrenchActionStateComponent.cpp">
+      <Filter>Rachet\Source\Component\Weapon\OmniWrench</Filter>
+    </ClCompile>
+    <ClCompile Include="Component\Weapon\OmniWrench\OmniWrenchDefaultComponent.cpp">
+      <Filter>Rachet\Source\Component\Weapon\OmniWrench</Filter>
+    </ClCompile>
+    <ClCompile Include="Component\Weapon\OmniWrench\OmniWrenchThrowedComponent.cpp">
+      <Filter>Rachet\Source\Component\Weapon\OmniWrench</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1484,20 +1496,11 @@
     <ClInclude Include="Component\Collision\Algolithm\EnemyOmniWrenchCollisionAlgolithm.h">
       <Filter>Rachet\Header\Component\Collision\Algolithm</Filter>
     </ClInclude>
-    <ClInclude Include="Component\Player\PlayerWeaponComponent.h">
+    <ClInclude Include="Component\Player\PlayerOmniWrenchComponent.h">
       <Filter>Rachet\Header\Component\Player</Filter>
-    </ClInclude>
-    <ClInclude Include="Component\Weapon\OmniWrenchThrowedComponent.h">
-      <Filter>Rachet\Header\Component\Weapon</Filter>
-    </ClInclude>
-    <ClInclude Include="Component\Weapon\OmniWrenchActionStateComponent.h">
-      <Filter>Rachet\Header\Component\Weapon</Filter>
     </ClInclude>
     <ClInclude Include="State\OmniWrenchActionStateDefine.h">
       <Filter>Rachet\Header\State</Filter>
-    </ClInclude>
-    <ClInclude Include="Component\Weapon\OmniWrenchDefaultComponent.h">
-      <Filter>Rachet\Header\Component\Weapon</Filter>
     </ClInclude>
     <ClInclude Include="Camera\FirstPersonCameraController.h">
       <Filter>Rachet\Header\Camera</Filter>
@@ -1591,6 +1594,21 @@
     </ClInclude>
     <ClInclude Include="Component\Player\PlayerThrowAttackSetComponent.h">
       <Filter>Rachet\Header\Component\Player\Action</Filter>
+    </ClInclude>
+    <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchActionStateComponent.h">
+      <Filter>Rachet\Header\Component\Weapon\OmniWrench</Filter>
+    </ClInclude>
+    <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchDefaultComponent.h">
+      <Filter>Rachet\Header\Component\Weapon\OmniWrench</Filter>
+    </ClInclude>
+    <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchThrowedComponent.h">
+      <Filter>Rachet\Header\Component\Weapon\OmniWrench</Filter>
+    </ClInclude>
+    <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchComponentInclude.h">
+      <Filter>Rachet\Header\Component\Weapon\OmniWrench</Filter>
+    </ClInclude>
+    <ClInclude Include="Factory\ComponentRegister\OmniWrenchComponentRegister.h">
+      <Filter>Rachet\Header\Factory\ComponentRegister</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
