+++ ClientProject/Mof/RachetNClankCopy/Resource/builder/player.json
@@ -44,7 +44,7 @@
     },
 
     {
-      "type": "PlayerWeaponComponent",
+      "type": "PlayerOmniWrenchComponent",
       "paramater": {
         "priority": 85
       }
