+++ ClientProject/Mof/RachetNClankCopy/Resource/scene_resource/game_scene.txt
@@ -6,6 +6,12 @@
 ../Resource/mesh/skeleton/skeleton.mom
 ../Resource/mesh/futuristic_transport_shuttle_animated/scene.mom
 
+../Resource/mesh/blaster_bullet/scene.mom
+../Resource/mesh/pyrocitor_bullet/scene.mom
+
+
+
+
 ../Resource/texture/icon/bomb_glove.png
 ../Resource/texture/icon/pyrocitor.png
 ../Resource/texture/icon/blaster.png
